Manually unbox Integer to int

+++ src/tuwien/auto/calimero/GroupAddress.java
@@ -65,7 +65,7 @@ public class GroupAddress extends KNXAddress
 	/**
 	 * Creates a KNX group address from a 16 Bit address value.
 	 * <p>
-	 * 
+	 *
 	 * @param address the address value in the range 0 <= value <= 0xFFFF
 	 */
 	public GroupAddress(final int address)
@@ -76,7 +76,7 @@ public class GroupAddress extends KNXAddress
 	/**
 	 * Creates a KNX group address from the 3-level notation main-, middle- and sub-group.
 	 * <p>
-	 * 
+	 *
 	 * @param mainGroup main group value, in the range 0 <= value <= 0x1F
 	 * @param middleGroup middle group value, in the range 0 <= value <= 0x7
 	 * @param subGroup sub group value, in the range 0 <= value <= 0xFF
@@ -89,7 +89,7 @@ public class GroupAddress extends KNXAddress
 	/**
 	 * Creates a KNX group address from the 2-level notation main- and sub-group.
 	 * <p>
-	 * 
+	 *
 	 * @param mainGroup main group value, in the range 0 <= value <= 0x1F
 	 * @param subGroup sub group value, in the range 0 <= value <= 0x7FF
 	 */
@@ -103,7 +103,7 @@ public class GroupAddress extends KNXAddress
 	 * <p>
 	 * The address is read out of the first 2 byte fields, while the address array itself
 	 * might be longer. The content of <code>address</code> is not modified.
-	 * 
+	 *
 	 * @param address the address byte array in big-endian format, with address.length > 1
 	 */
 	public GroupAddress(final byte[] address)
@@ -117,7 +117,7 @@ public class GroupAddress extends KNXAddress
 	 * The address string can either be formatted, e.g., "2/1/2", or the raw address, e.g., "4354".
 	 * A formatted address is a 2- or 3-level group address, the allowed separators are '.'
 	 * or '/', mutually exclusive.
-	 * 
+	 *
 	 * @param address string containing the KNX address
 	 * @throws KNXFormatException on unknown address type, wrong address syntax,
 	 *         group values out of range, or wrong separator used
@@ -127,7 +127,7 @@ public class GroupAddress extends KNXAddress
 		final String[] tokens = parse(address);
 		try {
 			if (tokens.length == 1)
-				init(Integer.decode(tokens[0]));
+				init(Integer.decode(tokens[0]).intValue());
 			else if (tokens.length == 2)
 				init(Byte.parseByte(tokens[0]), Short.parseShort(tokens[1]));
 			else if (tokens.length == 3)
@@ -147,7 +147,7 @@ public class GroupAddress extends KNXAddress
 	 * <p>
 	 * If the current XML element position is no start tag, the next element tag is read.
 	 * The KNX address element is then expected to be the current element in the reader.
-	 * 
+	 *
 	 * @param r a XML reader
 	 * @throws KNXMLException if the xml element is no KNX address or the
 	 *         address couldn't be read in correctly
@@ -162,7 +162,7 @@ public class GroupAddress extends KNXAddress
 	 * <p>
 	 * Note that, since this notation of levels only affects visual presentation and not
 	 * internal operation, this is a class setting.
-	 * 
+	 *
 	 * @param format3Level <code>true</code> for 3-level group format,
 	 *        <code>false</code> for 2-level group format
 	 */
@@ -174,7 +174,7 @@ public class GroupAddress extends KNXAddress
 	/**
 	 * Returns the current presentation of group addresses.
 	 * <p>
-	 * 
+	 *
 	 * @return <code>true</code> for 3-level formatting, <code>false</code> otherwise
 	 */
 	public static boolean is3LevelPresentation()
@@ -187,7 +187,7 @@ public class GroupAddress extends KNXAddress
 	 * <p>
 	 * The main group is equal for both the 2- and 3-level group address notation (see
 	 * class header specification).
-	 * 
+	 *
 	 * @return the main group value (5 most significant address bits)
 	 */
 	public final int getMainGroup()
@@ -200,7 +200,7 @@ public class GroupAddress extends KNXAddress
 	 * <p>
 	 * The middle group consists of 3 bits, starting with bit 8 to 10 in the address
 	 * field.
-	 * 
+	 *
 	 * @return the middle group value (3 bits)
 	 */
 	public final int getMiddleGroup()
@@ -212,7 +212,7 @@ public class GroupAddress extends KNXAddress
 	 * Returns the sub group value for the 3-level group notation.
 	 * <p>
 	 * The sub group consists of the low byte of the address field.
-	 * 
+	 *
 	 * @return the sub group value (8 least significant address bits)
 	 */
 	public final int getSubGroup8()
@@ -224,7 +224,7 @@ public class GroupAddress extends KNXAddress
 	 * Returns the sub group value for the 2-level group notation.
 	 * <p>
 	 * The sub group consists of the lower 11 bits in the address field.
-	 * 
+	 *
 	 * @return the sub group value (11 least significant address bits)
 	 */
 	public final int getSubGroup11()
@@ -245,7 +245,7 @@ public class GroupAddress extends KNXAddress
 	 * <p>
 	 * Depending on {@link #is3LevelPresentation()}, the address is formatted in 2- or
 	 * 3-level group notation.
-	 * 
+	 *
 	 * @return the address string
 	 */
 	public String toString()
@@ -259,7 +259,7 @@ public class GroupAddress extends KNXAddress
 	/**
 	 * Returns whether <code>obj</code> is equal to this KNX address (type).
 	 * <p>
-	 * 
+	 *
 	 * @param obj KNX address object
 	 * @return <code>true</code> iff <code>obj</code> is of this type and contains the
 	 *         same address (raw), <code>false</code> otherwise
